refactor: gameobject モジュールを object にリネームし、関連するルーティングを更新
